Improve compare_numeric()
New strategy where if inputs aren't the same length, we just use as.character(), otherwise we used the smallest number of digits needed to show the smallest difference. This should only error on the side of too many digits, and if there's some alignment that decreases the smallest difference.

Also improves handling of missing values

Fixes #43